Removed unnecessary files
- Removed the file from the lib folder - realistically I have no idea if
it is not need or really what it does at all. However, when I deleted it
the window and everything still ran perfectly fine, so I think deleting
it for now is fine, though if stuff down the line starts to go horribly
wrong, adding it back in may not be a terrible idea.