Removed unneeded catch2 library
I know a bit more about this now - I'm pretty sure this is a library used for testing, but I'm just gonna remove it since we have no idea how to use it anyway.
I think the problem I ran into before when I did this was I deleted the /lib/ folder as well meaning the Makefile couldn't see that it was empty.
But this file has been deleted for a while now and I haven't run into any problems as of yet, but we'll see.
I also added a placeholder file to the /lib/ dir so that it still gets pulled upon cloning so that the project still works.