Added delta time fix
This commit technically does not work because for some reason at the start of the program the framerate acts a bit weird.
It causes everything effected by gravity (player and enemies) to fall directly through the platforms and therefore go beneath the screen.
This then immediately fixes itself as the framerate stabalises.

I am not going to try to fix this because when there is a menu beforehand, it will allow time before the players and enemies are placed, meaning that issue should never occur.