Changed enemies and platforms to std::vector from basic C array
std::vector is dynamic so it can be freely added to, which will be important later on.
It's also much easier to iterate through.

I needed the vectors to be of the class pointers as sometimes (by that I mean all the time) the textures of the sprites would unload since the vector would keep trying to move the memory around which would break the reference between the texture and the sprite causing it to show a tan rectangle instead of the actual picture.